Refactor prediction query into service module
